+++ resources/views/tasks/show.blade.php
@@ -18,11 +18,12 @@
          <th>{{ $task->content }}</th>
       </tr>
    </table>
+    
+      {!! link_to_route('tasks.edit', 'このタスクを編集', ['id' => $task->id], ['class' => 'btn btn-light'])!!}
    
-   {!! link_to_route('tasks.edit', 'このタスクを編集', ['id' => $task->id], ['class' => 'btn btn-light'])!!}
-   
-   {!! Form::model($task, ['route' => ['tasks.destroy', $task->id], 'method' => 'delete']) !!}
-      {!! Form::submit('削除', ['class' => 'btn btn-danger']) !!}
-   {!! Form::close() !!}
-
+   @if (Auth::id() == $task->user_id)
+      {!! Form::model($task, ['route' => ['tasks.destroy', $task->id], 'method' => 'delete']) !!}
+         {!! Form::submit('削除', ['class' => 'btn btn-danger']) !!}
+      {!! Form::close() !!}
+   @endif
 @endsection
